[Java]: make sure to cast termLength to long so that 1GB term lengths compute correct log length.

+++ aeron-client/src/main/java/io/aeron/logbuffer/LogBufferDescriptor.java
@@ -602,7 +602,7 @@ public class LogBufferDescriptor
             return align((termLength * PARTITION_COUNT) + LOG_META_DATA_LENGTH, filePageSize);
         }
 
-        return (PARTITION_COUNT * termLength) + align(LOG_META_DATA_LENGTH, filePageSize);
+        return (PARTITION_COUNT * (long)termLength) + align(LOG_META_DATA_LENGTH, filePageSize);
     }
 
     /**
